backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.16.0
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==22.1
django-ratelimit==3.0.1
django-redis==5.2.0
djangorestframework==3.13.1
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==20.1.0
importlib-metadata==4.12.0
inflection==0.5.1
Markdown==3.3.7
mccabe==0.7.0
packaging==21.3
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0.1
redis==4.3.4
sentry-sdk==1.7.2
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
zipp==3.8.0
